+++ covalent_ui/result_webhook.py
@@ -19,6 +19,7 @@
 # Relief from the License may be granted by purchasing a commercial license.
 
 import json
+import os
 
 import requests
 
@@ -37,7 +38,9 @@ app_log = logger.app_log
 
 
 def get_ui_url(path):
-    baseUrl = f"http://localhost:{DEFAULT_PORT}"
+    port = 8080 if os.environ.get("PLATFORM") == "Docker" else DEFAULT_PORT
+    baseUrl = f"http://localhost:{port}"
+    app_log.debug(f"ui url is {baseUrl}{path}")
     return f"{baseUrl}{path}"
 
 
@@ -68,7 +71,7 @@ def send_update(result: Result) -> None:
         requests.post(get_ui_url(ui_server.WEBHOOK_PATH), data=result_update)
     except requests.exceptions.RequestException:
         # catch all requests-related exceptions
-        app_log.warning("Unable to send result update to UI server.")
+        app_log.exception("Unable to send result update to UI server.")
 
 
 def send_draw_request(lattice) -> None: